move components into more specific directories
