+++ src/app/instances/[name]/firewall/page.tsx
@@ -5,14 +5,14 @@ import axios from "axios";
 import React from "react";
 import { useInstanceContext } from "../InstanceContext";
 import { FirewallRule } from "@/types/firewall";
-import ErrorBanner from "@/app/components/ErrorBanner";
+import ErrorBanner from "@/app/instances/components/ErrorBanner";
 import { isValidDescription, isValidSourceIp, isInRangeCustomPort } from "@/utils/firewallValidation";
 import { COMMON_PORTS } from "@/utils/firewallConstants";
 import { Info } from "lucide-react";
 import { Trash2 } from "lucide-react";
 
 import { useNotificationsContext } from "@/app/NotificationContext";
-import SubmissionSpinner from "@/app/components/SubmissionSpinner";
+import SubmissionSpinner from "../../components/SubmissionSpinner";
 
 export default function FirewallPage() {
   const { instance } = useInstanceContext();
@@ -59,15 +59,15 @@ export default function FirewallPage() {
     setRules((prevRules) => {
       const updatedRules = [...prevRules];
       updatedRules[index] = { ...updatedRules[index], description: value };
-  
+
       const error = validateDescription(value);
       resetError(`Description must be 255 characters or fewer, and cannot contain the following characters: ^, ", ', %, &, <, >, |, \``);
-  
+
       if (error) addError(error);
       return updatedRules;
     });
   };
-  
+
   const handleSourceIpChange = (index: number, value: string) => {
     setRules((prevRules) => {
       const updatedRules = [...prevRules];
@@ -83,7 +83,7 @@ export default function FirewallPage() {
     resetError("Invalid Source IP format.");
     if (error) addError(error);
   };
-  
+
   const handleCustomPortsChange = (index: number, value: string) => {
     setRules((prevRules) => {
       const updatedRules = [...prevRules];
@@ -143,51 +143,51 @@ export default function FirewallPage() {
     }
     return null;
   }
-  
+
   const validateSourceIp = (sourceIp: string): string | null => {
     const inputError =
       "Invalid Source IP format. Use CIDR block notation, e.g. '192.168.0.0/24' or '10.0.0.1/32'.";
-      
+
     if (sourceIp.trim() !== "" && !isValidSourceIp(sourceIp)) {
       return inputError;
     }
-  
+
     return null;
   };
-  
+
   const validateCustomPorts = (customPorts: string, commonPorts: string[]): string[] => {
     const errors: string[] = [];
-  
+
     const portList = customPorts
       .split(",")
       .map((port) => port.trim())
       .filter((port) => port !== "");
-  
+
     const nonNumericPorts = portList.filter((port) => !/^\d+$/.test(port));
-  
+
     if (nonNumericPorts.length > 0) {
       errors.push("Custom ports must be a comma-separated list of numbers.");
       return errors;
     }
-  
+
     if (!isInRangeCustomPort(customPorts)) {
       errors.push("Ports must be between 1 and 65535.");
     }
-  
+
     const repeated = findCommonAndCustomPortOverlap(customPorts, commonPorts);
     if (repeated.length > 0) {
       errors.push("Port is already listed as a common port.");
     }
-  
+
     return errors;
   };
-  
+
   const findCommonAndCustomPortOverlap = (customPorts: string, commonPorts: string[]): string[] => {
     const customList = customPorts
       .split(",")
       .map((port) => port.trim())
       .filter((port) => port !== "");
-  
+
     const commonPortNumbers = COMMON_PORTS.filter(({ name }) => commonPorts.includes(name)).map((p) => p.port.toString());
     return customList.filter((port) => commonPortNumbers.includes(port));
   }
@@ -204,33 +204,33 @@ export default function FirewallPage() {
     });
 
     const validationErrors: string[] = [];
-  
+
     rules.forEach((rule) => {
       const descError = validateDescription(rule.description);
       const sourceIpError = validateSourceIp(rule.sourceIp);
       const customPortErrors = validateCustomPorts(rule.customPorts, rule.commonPorts);
-  
+
       if (descError) validationErrors.push(descError);
       if (sourceIpError) validationErrors.push(sourceIpError);
       validationErrors.push(...customPortErrors);
     });
-  
+
     setErrors(validationErrors);
-  
+
     if (validationErrors.length > 0) return;
-  
+
     try {
       const { data } = await axios.put(
         `/api/instances/${instance?.name}/firewall?region=${instance?.region}`,
         { rules }
       );
-  
+
       setRules(data);
     } catch (error) {
       console.error("Error saving rules:", error);
     }
   };
-  
+
   const handleReset = async () => {
     setIsLoading(true);
     try {
@@ -427,12 +427,12 @@ export default function FirewallPage() {
               disabled={errors.length > 0 || formPending()}
               className="font-heading1 bg-btn1 text-mainbg1 font-semibold px-4 py-2 rounded-sm hover:bg-btnhover1 cursor-pointer flex items-center justify-center hover:shadow-[0_0_10px_#87d9da] transition-all duration-200"
             >
-            {formPending() ? 
-              <span className="flex items-center gap-2">
+              {formPending() ?
+                <span className="flex items-center gap-2">
                   <SubmissionSpinner />
                   Saving ...
-              </span>
-              : "Save"}
+                </span>
+                : "Save"}
             </button>
           </div>
         </div>
