+++ src/app/instances/components/ErrorBanner.tsx
