+++ src/app/instances/components/InstanceDetails.tsx
