+++ src/app/instances/components/StorageDetails.tsx
