+++ src/app/instances/page.tsx
@@ -5,7 +5,7 @@ import axios from "axios";
 import { Trash2 } from "lucide-react";
 import { useEffect, useState } from "react";
 import { useNotificationsContext } from "../NotificationContext";
-import SubmissionSpinner from "../components/SubmissionSpinner";
+import SubmissionSpinner from "./components/SubmissionSpinner";
 
 interface Instance {
   state: string;
@@ -142,96 +142,94 @@ export default function Home() {
         <tbody className={isLoading ? "" : "animate-fade-in"}>
           {isLoading
             ? Array.from({ length: 3 }).map((_, idx) => (
-                <tr
-                  key={idx}
-                  className="bg-card border border-gray-500/30 animate-pulse"
-                >
-                  <td className="px-4 py-3">
-                    <div className="w-full h-4 bg-gray-600 rounded"></div>
-                  </td>
-                  <td className="px-4 py-3">
-                    <div className="w-full h-4 bg-gray-600 rounded"></div>
-                  </td>
-                  <td className="px-4 py-3">
-                    <div className="w-full h-4 bg-gray-600 rounded"></div>
-                  </td>
-                  <td className="px-4 py-3">
-                    <div className="w-full h-4 bg-gray-600 rounded"></div>
-                  </td>
-                  <td className="px-4 py-3 text-right">
-                    <div className="w-5 h-4 bg-gray-600 rounded ml-auto"></div>
-                  </td>
-                </tr>
-              ))
+              <tr
+                key={idx}
+                className="bg-card border border-gray-500/30 animate-pulse"
+              >
+                <td className="px-4 py-3">
+                  <div className="w-full h-4 bg-gray-600 rounded"></div>
+                </td>
+                <td className="px-4 py-3">
+                  <div className="w-full h-4 bg-gray-600 rounded"></div>
+                </td>
+                <td className="px-4 py-3">
+                  <div className="w-full h-4 bg-gray-600 rounded"></div>
+                </td>
+                <td className="px-4 py-3">
+                  <div className="w-full h-4 bg-gray-600 rounded"></div>
+                </td>
+                <td className="px-4 py-3 text-right">
+                  <div className="w-5 h-4 bg-gray-600 rounded ml-auto"></div>
+                </td>
+              </tr>
+            ))
             : instances.map((instance) => (
-                <tr
-                  key={instance.name}
-                  className="bg-card border border-gray-500/30"
-                >
-                  <td className="px-4 py-3 relative">
-                    {instance.state === "pending" ||
+              <tr
+                key={instance.name}
+                className="bg-card border border-gray-500/30"
+              >
+                <td className="px-4 py-3 relative">
+                  {instance.state === "pending" ||
                     instance.state === "shutting-down" ||
                     instance.state === "terminated" ||
                     instancePending(instance.name) ? (
-                      <span className="text-pagetext1 truncate block group cursor-not-allowed">
-                        {instance.name}
-                      </span>
-                    ) : (
-                      <Link
-                        href={`/instances/${instance.name}?region=${instance.region}`}
-                        className="text-pagetext1 hover:text-btnhover1 truncate block"
-                      >
-                        {instance.name}
-                      </Link>
-                    )}
-                  </td>
-                  <td className="px-4 py-3 text-pagetext1 truncate">
-                    {instance.id}
-                  </td>
-                  <td className="px-4 py-3 text-pagetext1">
-                    {instance.region}
-                  </td>
-                  <td
-                    className={`px-4 py-3 ${
-                      instance.state === "running"
-                        ? "text-btnhover1"
-                        : instance.state === "pending" ||
-                          instance.state === "initializing"
-                        ? "text-btn1 italic"
-                        : instance.state === "stopped" ||
-                          instance.state === "stopping"
+                    <span className="text-pagetext1 truncate block group cursor-not-allowed">
+                      {instance.name}
+                    </span>
+                  ) : (
+                    <Link
+                      href={`/instances/${instance.name}?region=${instance.region}`}
+                      className="text-pagetext1 hover:text-btnhover1 truncate block"
+                    >
+                      {instance.name}
+                    </Link>
+                  )}
+                </td>
+                <td className="px-4 py-3 text-pagetext1 truncate">
+                  {instance.id}
+                </td>
+                <td className="px-4 py-3 text-pagetext1">
+                  {instance.region}
+                </td>
+                <td
+                  className={`px-4 py-3 ${instance.state === "running"
+                    ? "text-btnhover1"
+                    : instance.state === "pending" ||
+                      instance.state === "initializing"
+                      ? "text-btn1 italic"
+                      : instance.state === "stopped" ||
+                        instance.state === "stopping"
                         ? "text-red-300"
                         : instance.state === "shutting-down" ||
                           instance.state === "terminated"
-                        ? "text-pagetext1 italic"
-                        : ""
+                          ? "text-pagetext1 italic"
+                          : ""
                     }`}
-                  >
-                    {instance.state}
-                  </td>
-                  <td className="px-4 py-3 text-right">
-                    <button
-                      onClick={() => openDeleteModal(instance)}
-                      className={`
+                >
+                  {instance.state}
+                </td>
+                <td className="px-4 py-3 text-right">
+                  <button
+                    onClick={() => openDeleteModal(instance)}
+                    className={`
                   text-gray-400 
-                  ${
-                    instance.state === "pending" ||
-                    instance.state === "shutting-down"
-                      ? "cursor-not-allowed"
-                      : "hover:text-btnhover1 hover:shadow-btnhover1"
-                  }
-                `}
-                      aria-label="Delete instance"
-                      disabled={
-                        instance.state === "pending" ||
+                  ${instance.state === "pending" ||
                         instance.state === "shutting-down"
+                        ? "cursor-not-allowed"
+                        : "hover:text-btnhover1 hover:shadow-btnhover1"
                       }
-                    >
-                      <Trash2 size={20} />
-                    </button>
-                  </td>
-                </tr>
-              ))}
+                `}
+                    aria-label="Delete instance"
+                    disabled={
+                      instance.state === "pending" ||
+                      instance.state === "shutting-down"
+                    }
+                  >
+                    <Trash2 size={20} />
+                  </button>
+                </td>
+              </tr>
+            ))}
         </tbody>
       </table>
       {!isLoading && instances.length === 0 && (
